requirements.txt
absl-py==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.4.0
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
certifi==2025.6.15
chardet==4.0.0
colorama==0.4.6
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
eagerpy==0.30.0
exceptiongroup==1.3.0
executing==2.2.0
filelock==3.18.0
fonttools==4.58.5
foolbox==3.3.4
frozenlist==1.7.0
fsspec==2025.5.1
fvcore==0.1.5.post20221221
gitdb==4.0.12
GitPython==3.1.44
idna==2.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
intel-openmp==2021.4.0
iopath==0.1.10
ipykernel==6.29.5
ipython==8.18.1
ipywidgets==8.1.7
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.1
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyterlab_widgets==3.0.15
kiwisolver==1.4.7
kornia==0.8.1
kornia_rs==0.1.9
lightning-utilities==0.14.3
MarkupSafe==3.0.2
matplotlib==3.9.4
matplotlib-inline==0.1.7
mkl==2021.4.0
mpmath==1.3.0
multidict==6.6.3
nest-asyncio==1.6.0
networkx==3.2.1
ninja==1.11.1.4
numpy==1.26.3
packaging==25.0
pandas==2.3.0
parso==0.8.4
pastel==0.2.1
pillow==11.3.0
platformdirs==4.3.8
poethepoet==0.10.0
portalocker==3.2.0
POT==0.9.5
prompt_toolkit==3.0.51
propcache==0.3.2
psutil==7.0.0
pure_eval==0.2.3
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytorch-lightning==2.5.2
pytz==2025.2
pywin32==310
PyYAML==6.0.2
pyzmq==27.0.0
requests==2.25.1
scikit-learn==1.6.1
scipy==1.13.1
six==1.17.0
smmap==5.0.2
stack-data==0.6.3
sympy==1.14.0
tabulate==0.9.0
tbb==2021.13.1
termcolor==3.1.0
threadpoolctl==3.6.0
tomlkit==0.13.3
torchattacks==3.5.1
torchcde==0.2.5
torchcfm==1.0.7
torchdiffeq==0.2.5
torchdyn==1.0.6
torchmetrics==1.7.4
torchsde==0.2.6
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
trampoline==0.1.2
typing_extensions==4.14.1
tzdata==2025.2
urllib3==1.26.20
wcwidth==0.2.13
widgetsnbextension==4.0.14
yacs==0.1.8
yarl==1.20.1
zipp==3.23.0
